PLAT-41: Rename the platformui owa route urls
