+++ app/js/components/App.jsx
@@ -1,4 +1,4 @@
-/* * This Source Code Form is subject to the terms of the Mozilla Public License,
+/** This Source Code Form is subject to the terms of the Mozilla Public License,
  * v. 2.0. If a copy of the MPL was not distributed with this file, You can
  * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
  * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
@@ -7,18 +7,10 @@
  * graphic logo is a trademark of OpenMRS Inc.
  */
 import React from 'react';
-import Login from './Login';
+import Dashboard from './Dashboard';
 
-export default class App extends React.Component {
-  render() {
-    return (
-      <div className="container">
-        <div className="row">
-          <div className="col-md-6 offset-md-5 main-container">
-            <Login />
-          </div>
-        </div>
-      </div>
-    )
-  }
-}
+const App = () => (
+    <Dashboard />
+);
+
+export default App;
